clean up warnings for bench under vs2019

+++ performance-tests/bench_2/common/util.cpp
@@ -27,7 +27,7 @@ int get_option_argument_int(int& i, int argc, ACE_TCHAR* argv[])
 {
   int value;
   try {
-    value = std::stoll(get_option_argument(i, argc, argv));
+    value = static_cast<int>(std::stoll(get_option_argument(i, argc, argv)));
   } catch (const std::exception&) {
     std::cerr << "Option " << ACE_TEXT_ALWAYS_CHAR(argv[i]) << " requires an argument that's a valid number" << std::endl;
     throw 1;
@@ -39,7 +39,7 @@ unsigned get_option_argument_uint(int& i, int argc, ACE_TCHAR* argv[])
 {
   unsigned value;
   try {
-    value = std::stoull(get_option_argument(i, argc, argv));
+    value = static_cast<unsigned>(std::stoull(get_option_argument(i, argc, argv)));
   } catch (const std::exception&) {
     std::cerr << "Option " << ACE_TEXT_ALWAYS_CHAR(argv[i])
       << " requires an argument that's a valid positive number" << std::endl;
